refactor : 4단계 코드리뷰 내용 적용 - 자동차 이름 경계값 테스트 추가 - 우승자 조회 로직을 Race.kt에서 Cars.kt로 변경
