+++ src/main/kotlin/racingcar/race/Race.kt
@@ -13,14 +13,4 @@ class Race(
                 { index, car -> car.move(randomNumbers.randomNumbers[index]) },
         )
     }
-
-    // 우승자 조회
-    fun getWinners(): Cars {
-        // 최장 이동 거리를 찾는다.
-        val maxDistance = cars.list.maxBy { it.distance }.distance
-
-        // 최장 이동거리에 해당하는 Car 목록을 반환한다.
-        val winners = cars.list.filter { it.distance == maxDistance }
-        return Cars(winners)
-    }
 }
