+++ src/test/kotlin/racingcar/car/CarTest.kt
@@ -14,7 +14,7 @@ class CarTest {
 
     @Test
     fun `자동차 이름은 5자를 초과하는 경우 예외를 던진다`() {
-        assertThatThrownBy { Car(name = "testName") }
+        assertThatThrownBy { Car(name = "test12") }
             .isInstanceOf(IllegalArgumentException::class.java)
     }
